# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  seeker: "seeker",
  employer: "employer",
  admin: "admin",
} as const;

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  online: "online",
  field: "field",
  hybrid: "hybrid",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface PlatformStats {
  totalJobs: number;
  totalEmployers: number;
  totalSeekers: number;
  totalApplications: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PublicJob {
  id: number;
  title: string;
  description: string;
  type: JobType;
  category: string;
  employerName: string;
  /** @nullable */
  employerLocation?: string | null;
  /** @nullable */
  deadline?: string | null;
  createdAt: string;
}

export interface PublicJobDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  employerName: string;
  /** @nullable */
  employerLocation?: string | null;
  /** @nullable */
  employerBio?: string | null;
  /** @nullable */
  deadline?: string | null;
  isOpen: boolean;
  createdAt: string;
  savedByMe: boolean;
  appliedByMe: boolean;
}

export interface PublicJobListResponse {
  items: PublicJob[];
  total: number;
}

export interface CurrentUser {
  id: string;
  name: string;
  email: string;
  role: Role;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
  isActive: boolean;
  onboarded: boolean;
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
}

export type SetRoleBodyRole =
  (typeof SetRoleBodyRole)[keyof typeof SetRoleBodyRole];

export const SetRoleBodyRole = {
  seeker: "seeker",
  employer: "employer",
} as const;

export interface SetRoleBody {
  role: SetRoleBodyRole;
}

export interface CreateApplicationBody {
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
}

export interface MyApplication {
  id: number;
  jobId: number;
  jobTitle: string;
  employerName: string;
  status: ApplicationStatus;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
  createdAt: string;
}

export interface SeekerDashboard {
  totalApplications: number;
  pendingApplications: number;
  acceptedApplications: number;
  rejectedApplications: number;
  savedJobsCount: number;
  recentApplications: MyApplication[];
  recommendedJobs: PublicJob[];
}

export interface CreateJobBody {
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  /** @nullable */
  deadline?: string | null;
}

export interface UpdateJobBody {
  title?: string;
  description?: string;
  /** @nullable */
  requirements?: string | null;
  type?: JobType;
  category?: string;
  contactInfo?: string;
  /** @nullable */
  deadline?: string | null;
}

export interface EmployerJob {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  status: JobStatus;
  /** @nullable */
  rejectionReason?: string | null;
  isOpen: boolean;
  /** @nullable */
  deadline?: string | null;
  applicationsCount: number;
  unseenApplicationsCount: number;
  createdAt: string;
}

export interface EmployerApplication {
  id: number;
  jobId: number;
  applicantName: string;
  applicantEmail: string;
  /** @nullable */
  applicantPhone?: string | null;
  /** @nullable */
  applicantLocation?: string | null;
  /** @nullable */
  applicantBio?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
  status: ApplicationStatus;
  seenByEmployer: boolean;
  createdAt: string;
}

export type UpdateApplicationBodyStatus =
  (typeof UpdateApplicationBodyStatus)[keyof typeof UpdateApplicationBodyStatus];

export const UpdateApplicationBodyStatus = {
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface UpdateApplicationBody {
  status: UpdateApplicationBodyStatus;
}

export interface EmployerDashboard {
  totalJobs: number;
  approvedJobs: number;
  pendingJobs: number;
  rejectedJobs: number;
  totalApplications: number;
  unseenApplications: number;
  recentApplications: EmployerApplication[];
  topJobs: EmployerJob[];
}

export interface AdminJob {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  status: JobStatus;
  /** @nullable */
  rejectionReason?: string | null;
  employerId: string;
  employerName: string;
  employerEmail: string;
  createdAt: string;
}

export interface RejectJobInput {
  reason: string;
}

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  role: Role;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  isActive: boolean;
  jobsCount: number;
  applicationsCount: number;
  createdAt: string;
}

export interface AdminDashboard {
  totalUsers: number;
  totalSeekers: number;
  totalEmployers: number;
  totalJobs: number;
  pendingJobs: number;
  approvedJobs: number;
  rejectedJobs: number;
  totalApplications: number;
  recentPendingJobs: AdminJob[];
}

export interface RequestUploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResult {
  uploadURL: string;
  objectPath: string;
}

export type ListJobsParams = {
  search?: string;
  type?: ListJobsType;
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListJobsType = (typeof ListJobsType)[keyof typeof ListJobsType];

export const ListJobsType = {
  online: "online",
  field: "field",
  hybrid: "hybrid",
} as const;

export type ListAdminJobsParams = {
  status?: ListAdminJobsStatus;
};

export type ListAdminJobsStatus =
  (typeof ListAdminJobsStatus)[keyof typeof ListAdminJobsStatus];

export const ListAdminJobsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ListAdminUsersParams = {
  role?: ListAdminUsersRole;
};

export type ListAdminUsersRole =
  (typeof ListAdminUsersRole)[keyof typeof ListAdminUsersRole];

export const ListAdminUsersRole = {
  seeker: "seeker",
  employer: "employer",
  admin: "admin",
} as const;