# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fursa Platform API - Digital Employment Platform for Gaza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/jobs:
    get:
      operationId: listAdminJobs
      tags: [admin]
      summary: List jobs for admin moderation
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminJob"
  /admin/jobs/{id}/approve:
    post:
      operationId: approveJob
      tags: [admin]
      summary: Approve a pending job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJob"
  /admin/jobs/{id}/reject:
    post:
      operationId: rejectJob
      tags: [admin]
      summary: Reject a job with a reason
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectJobInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminJob"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [seeker, employer, admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/toggle-active:
    post:
      operationId: toggleUserActive
      tags: [admin]
      summary: Activate / deactivate a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
components:
  schemas:
...
    Role:
      type: string
      enum: [seeker, employer, admin]
    JobType:
      type: string
      enum: [online, field, hybrid]
    JobStatus:
      type: string
      enum: [pending, approved, rejected]
...
    AdminJob:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        contactInfo:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        rejectionReason:
          type: ["string", "null"]
        employerId:
          type: string
        employerName:
          type: string
        employerEmail:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - category
        - contactInfo
        - status
        - employerId
        - employerName
        - employerEmail
        - createdAt
    RejectJobInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    AdminUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
        jobsCount:
          type: integer
        applicationsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - jobsCount
        - applicationsCount
        - createdAt
...