# lib/api-zod/src/generated/types/adminJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from "./jobStatus";
import type { JobType } from "./jobType";

export interface AdminJob {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  type: JobType;
  category: string;
  contactInfo: string;
  status: JobStatus;
  /** @nullable */
  rejectionReason?: string | null;
  employerId: string;
  employerName: string;
  employerEmail: string;
  createdAt: Date;
}