# lib/api-zod/src/generated/types/rejectJobInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */

export interface RejectJobInput {
  reason: string;
}